server/requirements.txt
blinker
cffi
click
Flask
Flask-Admin
Flask-Login
Flask-SQLAlchemy
Flask-WTF
greenlet
gunicorn
itsdangerous
Jinja2
MarkupSafe
mysql-client
mysql-connector-python
mysqlclient
packaging
passlib
pillow
pycparser
pyheif
SQLAlchemy
typing_extensions
Werkzeug
WTForms
